Add interactivity via express cookies
Stores form values posted to the app as cookies, making the
values in the cookies available in templates that are rendered by the
catch-all get block in `server.js`.

Any default values that should be used if data is missing
or blank can be provided in `app/defaults.js` and any changes to the
rendering can be added to `app/presenters.js`.