Update markup for related links
Use the <aside> element to wrap this related content,
with an ARIA landmark role of 'complementary'.

Wrap the list of related links in a parent <nav> wrapper, with a role of 'navigation'.

Ensure that the title above the navigation is associated with the navigation using aria-labelledby, as
recommended in the style guide.

Update classname to match the 'related items' govuk component.