Add napa to dependencies to keep the govuk template up to date
This provides a way to keep the jinja version of the govuk template up
to date.

As npm won’t install a tarball without a package.json file, use napa to
do this.

Until a package.json file is added to the jinja version of the govuk
template use napa to install the govuk_template to node_modules.