Prevent asking users to authenticate twice
This solves the problem with prototypes asking for username/password
twice.

The problem is caused by the ordering in the middleware.

When a user types in the URL for a prototype like
http://govuk-tagging.herokuapp.com they're asked for a
username/password first (via `utils.basicAuth`).

After filling that in they'll be able to "proceed" to the next step,
which redirects them to the `https://` version (via
`utils.forceHttps`).

Because the auth isn't shared between `http` and `https` version, they
are not authenticated anymore and have to enter their username/password
again.

Validated with https://test-prevent-double-auth.herokuapp.com/