The PaaS team are conducting user research with the prototype kit.
We've found that the error message when basic auth hasn't been set
isn't as helpful as it could be. We've not yet added a PaaS-specific
error link, but hope to do that later once our own docs are ready for
release. We would dearly like to test the new version of this error
text with users in Newcastle next week.

This change also addresses a mis-entered URL in the error message where
a user hasn’t set up authentication around their prototype.

We also found performance improvements by copying .gitignore to
.cfignore - this prevented node.js dependencies being uploaded to PaaS
by the user and then downloaded as part of the buildpack process on
PaaS. Staging times were approximately halved for an unmodified
prototyping kit as a result. This change should only affect Cloud
Foundry users.