Restart the app if environment variables change
Could be confusing why your changes aren’t having any effect.

Doing this using the existing file-extension-based config doesn’t work
because nodemon doesn’t undestand that `.env` is a file with no name,
and an extension of `env`. So I’ve removed the file-extension-based
stuff entirely and rewritten it to use globbing.

Tested locally that:
- changing `server.js` or `.env` causes the app to restart
- changing a file inside `node_modules` _doesn’t_ cause the app to
  restart