Fix JS error in Safari’s Private Browsing mode
Safari in Private Browsing mode sets a session storage quota of 0 which means that attempts to write to it result in a QuotaExceededError being thrown, breaking JavaScript on the rest of the page.

The existing code is flawed in that it sets `prototypeWarning` to true but tests for 'false', so the warning is already shown on every page anyway.

As such, this simplifies it to remove the use of session storage entirely.