Support deprecated check-your-answers table styles
To ensure we support existing prototypes using the old check your
answers pattern - where using a table was recommended, add the styles
back in - but mark these as deprecated.

As the new styles are namespaced using .govuk- these are opt-in, so we
can release without making a breaking change.

In future, we can remove .check-your-answers, the next time we make a
breaking change to the prototype kit.