Integrate session name/secret into `.env`
In order to keep sessions secure, we need to use a secret for encryption. This cannot be kept in the repo, so we set it in the `.env` file and check for it when setting up session storage. If either the name or secret are missing, the app will exit with an error.