Generate name/secret for file-based sessions
If the designer/developer chooses to use file-based session storage, the app will generate a random name and secret using the crypto library. Once generated, these are stored in hidden files to be read from when the app restarts.

This commit also removes the use of the `.env` file for storing the name/secret provided by the designer/developer, in favour of auto-generation