Update dotenv from 4.0.0 to 6.0.0
dotenv 6.0 drops support for node v4, but this is fine because we already require node v6.

Other breaking changes are:
- default path is now path.resolve(process.cwd(), '.env')
- does not write over keys already in process.env if the key has a falsy value

Neither of these should affect the kit.

Tested by adding `FOO=bar` to the `.env` file, and ensuring that it is available within the app by adding a `console.log(process.env.FOO)` and checking for 'bar' in the console output.