Update marked from 0.3.6 to 0.4.0
Fixes a number of security vulnerabilities, fixes a load of bugs and adds a number of new features.

Some of these are listed as breaking changes, but only because they fix bugs so the output would be different.

Tested by going through the documentation (which is written in markdown) and checking that it looks OK.