Re-enable package-lock.json
Package lock files are a representation of the dependency tree which means that anyone running `npm install` is guaranteed to install exactly the same dependencies as specified in the lock file – including sub-dependencies, which are not mentioned in the package.json file at all.

We disabled this feature in #433 because it was causing issues for users. At the time, the package-lock.json file was not being committed, which means everyone who cloned the repo and ran npm install was generating the package-lock file for the first time. Because of this, we were confusing users by npm giving them  “created a lockfile as package-lock.json. You should commit this file” notices, whilst not actually getting any of the benefits.

Additionally, Node 5 which introduced the package-lock functionality had some issues which meant that the generated package-lock files were not deterministic – so depending on who ran `npm install` the file would change. Most of these issues appear to have been resolved in newer versions of Node. We have been using the package-lock functionality on our other repos (such as Frontend) without issue.

This therefore re-enables the generation of package-lock files and commits the generate package-lock.json file to source control.

It also overrides the default install command (`npm install`) to use `npm install --no-optional` to avoid an issue with Travis and the `fs-events` package – see https://github.com/npm/npm/issues/14042