Adds Extension Framework for departments to use their own frontend system
As a developer/Designer, I would like to use my departments frontend system
alongside govuk-frontend. So that I can prototype using the same components
and patterns that will be used in production.

This should not be a breaking change for any existing kits.
This is a soft release where GOV.UK Design System team would like HMRC
to test before releasing this more widely.

A proper release with documentation etc. will be done once this has been
battle tested with HMRC.

Co-authored-by: Alistair Laing <alistair.laing@digital.cabinet-office.gov.uk>
Co-authored-by: Hanna Laakso <hanna.laakso@digital.cabinet-office.gov.uk>