Rewrite instructions to add more context
The previous instructions were a very bare-bones draft which assumed a
lot of context.

This commit attempts to rewrite them for someone without knowledge of
what Notify is, or how an API works.

Hopefully this is enough so that someone who is confident using the
prototype kit can integrate their prototype with Notify.