Add a manifest file for GOV.UK PaaS users
This makes the process for deploying the prototype kit to GOV.UK PaaS
for new users slightly nicer.

Previously you had to set the USERNAME and PASSWORD variables
separately, like so:

```
$ cf push your-prototype-name
...
$ cf set-env your-prototype-name USERNAME myusername
...
$ cf set-env your-prototype-name PASSWORD mysecretpassword
...
$ cf restart your-prototype-name
...
```

With the manifest in place it becomes:

```
$ cf push your-prototype-name --var username=myusername --var password=mysecretpassword
```

If you try and run `cf push` without the `--var ...` flags it will fail
with:

```
...
Expected to find variables: password, username
...
```

This could be a breaking change for some users of the prototype kit, but
I think the usability is sufficiently better that it's worth doing.

I think the prototype kit currently has guidance for deploying to
Heroku, but not PaaS. I think it's best to discuss whether we should
update that guidance separately - this commit is just concerned with
making it easier to use the PaaS.