Test on on macOS and Windows and both Node 10 / 8
Update the Travis config to build on both macOS and Windows as well as Linux, and against both Node 8 and LTS (currently 10)

Travis combines the OS and Node JS versions, so both Node versions will be tested on every operating system.

There is a known issue with Windows builds on Travis where the build fails if you have encrypted variables.

Splitting the deploy step (which is the step that requires the encrypted variable) into its own job means that the encrypted variables are not present when testing on Windows.