Add new deploy key for Travis
In order to make the tests pass on Windows, we need to remove any encrypted variables from the Travis repository settings and specify them in the repo config file instead.

This adds a new deploy key encrypted using an encrypted env variable DEPLOY_KEY and updates the release script to use it.

Co-authored-by: Hanna Laakso <hanna.laakso@digital.cabinet-office.gov.uk>