Update autoStoreData function to use Keypather library to allow referencing of session data using dot notation instead of array notation. This means you can be more consistent with way you reference values in the template and in the app code (routes files).
eg. {{ data.participant.profile.name }}

Data in the session can now be set with a parameter passed in the
url, also using dot notation.

eg. /question?participant.profile.name=stef

This data will be added to the session data, or overwrite existing
values. You may not want to merge this data into the session data
but still access the values in the template. To do this, you can
use the data.query value in the template to access the url parameters.
Include the additional parameter '&persist=No'  in the url. This will
prevent the values being merged into session.data, and subsequent
pages in the journey will not retain this value.

eg. /question?participant.profile.name=stef&persist=No

This would let you access the value in the template with:

{{ query.participant.profile.name }}

Add code to check for parameters passed in the query which have string
values wrapped in square brackets. It treats the comma as a delimiter
and turns the value into an array, attaching it to the session at the
path indicated by the parameter key. This is good for passing a
query in the url which contrains a list of checkboxes which should
be set when the page loads.

eg. /question?participant.profile.faveColors=[yellow,burgundy,pink]