Add explicit user group to Dockerfile
Due to kubernetes@445393fdcefa6d0354b7ce32a2304a7765fbd305 one must use
numeric user groups if you want to run a container as non-root on
Kubernetes.

This commit ensures the container runs as a non-root user with an
explicit numeric ID, so that it can run on a Kubernetes cluster that
forces MustRunAsNonRoot.