Turn off npm default auditing
We have decide to avoid showing users of the prototype kit audit messages.

This is because a lot of the low level issues are not easy to fix, and will cause unnecessary worry.

We will prioritise the security alerts we recieve from GitHub's monitoring.

See https://github.com/alphagov/govuk-prototype-kit/issues/699 for the full discussion.