Add config for prompting if default port unavailable
Set a config value for asking whether to try another port if the default is unavailable.

If set to true the behaviour remains the same.

If set to false the prototype will automatically try the next available port.

For kit users regularly running multiple prototypes locally this will avoid needing to interact with the prompt each time.