Remove unneeded settings import
As all the settings are `!default`, the expected way to override settings is to define them before doing the import.

Importing the settings _generally_ works, because we just end up redefining them before they get used. However, it does break some things – specifically, where settings rely on each other (such as the legacy mode settings, or $govuk-assets-path which is relied on by $govuk-images-path and $govuk-fonts-path).

This doesn't work correctly because e.g. $govuk-images-path gets defined before $govuk-assets-path gets overridden, which means it uses the _old_ assets path.

This caught out a user who was using this as a template, and was overriding $govuk-assets-path but not seeing the expected changes represented in $govuk-images-path.