Import only colour palette and applied colour
In d901c7f we removed an import of govuk-frontend/settings/all because it caused issues for users overriding settings that relied on each other.

However, this broke the unbranded stylesheet as it refers to `$govuk-body-background-colour` which was then undefined.

Importing only the colour palette and applied colour settings solves this problem. We also document the fact that if a user (for whatever reason) wants to enable compatibility mode or use the legacy palette then they should do that before the import statements.