Remove usage data tracking
We are no longer using the data that is tracked, so in line withe GDPR principle around data minimisation we can remove this.

We can use existing data to understand trends but I believe there is no good reason to be tracking this at the present time.

We have started testing in multiple operating systems, and will need to support latest version of Node.js, we do not need this data to make those decisions.

As a result of this we remove a step from first time users, build more trust and make it easier to run the kit in environment such as docker.

Data minimisation: https://ico.org.uk/for-organisations/guide-to-data-protection/guide-to-the-general-data-protection-regulation-gdpr/principles/data-minimisation/