Update Heroku stack to Heroku-18
The current Cedar-14 stack is end-of-life (EOL). After November 2, apps on Cedar-14 will continue to run, but may not build or receive security updates.

Changing the stack definition in app.json only changes the stack for new review apps. Once this has been tested, we’ll also need to change the stack for the current running ‘production’ app using the Heroku CLI or the Heroku Dashboard.

https://devcenter.heroku.com/articles/upgrading-to-the-latest-stack