Move the internal instructions on into internal_docs folder
The instructions for how to create and release a new version of the kit are hidden away in CONTRIBUTING.md.
This makes the instructions difficult to find and in a previous release we have forgotten to create the actual tag and release after bumping versions.
We also wouldn't expect an external contributor to release the kit, so CONTRIBUTING.md probably isn't the best place for this documentation.

This commit creates a new `internal_docs` folder and moves `linting.md` and the releasing docs into this folder so they're separate from documentation for prototype kit users