Point to latest release if not in ‘promo mode’
Now that we’re relying on the version defined in package.json, it will only ever link to the same version as used by the copy of the kit, not the actual latest version. This is fine for the docs site on Heroku, which should always be running the latest version, but could cause issues if someone tries to download the latest version by using the docs running locally in their kit.

When the kit is not running in ‘promo mode’ (as used by the live docs site), just point the user to the latest release page on GitHub.