Fix step 1 in branching tutorial page
Step 1 of tbe branching page in the tutorial tells the user to copy a
file called `template-content-page.html`, however no such file exists in
the prototype kit. I think it is asking the user to copy the file
`docs/views/templates/content.html`, so this commit updates the step to
make that clear.