Don’t swallow errors from session-data-defaults.js
If the code generating session-data-defaults is fairly complex, spotting and fixing those errors is a common problem, and is more than fixing simple syntax issues.

This change still allows the app to run, but will show the original error in the console alongside the warning about loading the defaults. By showing the full error it also highlights where there's a problem – the single line "Could not load" message can be easily missed in a busy log.

Showing the error in the log allows a developer to debug and fix the problem.