Upgrade notify client to version 5.1.0
No changes needed to the prototype documentation but this will mean that when a
user integrates their code with Notify and needs to look at the Notify
documentation, that the version of the Notify client they have installed
is the one used for the Notify documentation. We don't publish older
versions of the docs on our main website, although it will still be
available in git.