Downgrade lockfile to version 1
We think we still have users running Node version 12 and 14, which
install with npm@6 [[1], [2]]. npm@6 does understand lockfiles using
lockfileVersion@2, but will downgrade it automatically, resulting in
lockfile churn. If there are teams working on a prototype where some
members are using npm@6 and some are using npm@7 or up, this can get
really annoying, and result in a poor user experience.

npm@8 (distributed with Node v16 [[3]]) can still works with
lockfileVersion@1, so this commit changes the npm config and creates a
new (old) `package-lock.json` that will result in less lockfile churn [[4]].

In future when we drop support for Node v12 we can revisit this
decision, hopefully by then Node v14 will come with a new version of npm.

[1]: https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V12.md
[2]: https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V14.md
[3]: https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V16.md
[4]: https://docs.npmjs.com/cli/v7/using-npm/config#lockfile-version