Add bash script update.sh to simplify updating existing prototypes
We want to make it easier for users to update existing prototypes to use
the latest code from GitHub; the current method is entirely manual and
difficult for even developers to follow.

This is especially important for the (currently unreleased) changes to
add a package lockfile; for existing installations to be able to use a
package lockfile some hidden files need to be updated, which is
difficult to do manually.

This commit adds a small bash script to simplify some (but not all) of
the steps required to update an existing prototype. It covers steps 1,
2, (not 3), 4, 5, and 8 in the existing process [[1]]. The hypothesis is
that it will be easier for users to run the script than do these steps
manually, and easier to write a script that is robust and can do these
steps without breaking. Note that we don't include step 3, make a backup
copy of your prototype folder, because we felt that it is more useful
for the user to do this so they know where the backup is, and it reduces
the risk of the script somehow messing up the backup.

The script is currently quite 'chatty'; the theory behind this is that
it will be helpful for our users to see what the script is doing, and it
will also be helpful for us if we need to debug problems. It also logs
the final step of copying files from the new folder to the old in a file
`update/update.log`, again to aid with debugging.

We also tried to make sure that the script was relatively safe; it will
not run if the current working directory does not contain a prototype,
it avoids running commands twice (if for instance the script has already
been run), and it will exit immediately if any commands result in an
error.

---

Note to Linux users: this script is currently targeted at users of the
latest versions of macOS and Git for Windows, and uses commands which
are included as standard in those systems but are not included as
standard on Linux. You may need to ensure that curl and unzip are
installed on your system before the script will run correctly.

[1]: https://github.com/alphagov/govuk-prototype-kit/blob/ca3561625c03e1bbbb9a8b588bea5e5336bb14f8/docs/documentation/updating-the-kit.md

Co-authored-by: Joe Lanman <joe.lanman@digital.cabinet-office.gov.uk>
Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>