update.sh: Add CLEAN option
We want the update script to be idompotent (i.e. running it twice is not
a problem), and we also want to be able to test it without it
downloading new files, extracting new files etc.

So, to make the script more robust for users we remove any files from a
previous update, unless they supply the `CLEAN=0` option (via an
envvar). For most users this is not useful, but for testing it is very
useful.