update.sh: Add logging
We want stdout and stderr to be logged to update.log for post-mortem
debugging, but also want to be able to show an error message to the
user.

This commit acheives that in a fairly complicated way. As the comment
notes, the method was found mainly by trial and error, but some help was
found in the bash manual [[1]], an IBM manual on POSIX redirections
[[2]], and a bash cookbook page on swapping STDERR and STDOUT [[3]].

The idea to use a trap came from an Aaron Maxwell blog post [[4]], the
bash manual explains in a bit more detail about the ERR signal [[5]] and
how it is sent in the same conditions that are used by the `-e` shell
option.

[1]: https://www.gnu.org/software/bash/manual/html_node/Redirections.html
[2]: https://www.ibm.com/docs/en/aix/7.1?topic=ks-input-output-redirection-in-korn-shell-posix-shell
[3]: https://www.oreilly.com/library/view/bash-cookbook/0596526784/ch02s20.html
[4]: http://redsymbol.net/articles/bash-exit-traps/
[5]: https://www.gnu.org/savannah-checkouts/gnu/bash/manual/bash.html#index-trap