update.sh: Major refactor
Refactors update.sh to be composed of functions, so that it can be
tested easier Use technique from [[1]] (similar to how Python does it)
to create a script that can be both sourced (without being run) and
sourced.

The script is probably reaching critical mass of unmaintainability now,
hopefully this refactor will enable some readability improvements,
before we need to go full Node.js.

[1]: https://stackoverflow.com/a/29458276