update.sh: Extract update archive directly into update folder
Make it easier for the user by removing some of the directory heirarchy.

Note that this would be easier to do in bash if we used tar, which has a
`--strip-components` argument for doing exactly this operation, however
getting the latest release tarball from GitHub is slightly trickier than
getting the latest zip archive, because we don't have the same neat
redirect in place on Heroku.