update.sh: Fix running as piped script
When bash is reading script from stdin BASH_SOURCE is unset, so we make
$0 the default for those circumstances. This is fine because we only
really care about the if-main check when testing. This caught us out in
user testing, thankfully the `-u` shell option meant it was easy to
see what was happening and find the Stack Overflow post with the answer
to what was going on [[1]]. Thanks unofficial bash strict mode! [[2]].

[1]: https://stackoverflow.com/a/68359914
[2]: http://redsymbol.net/articles/unofficial-bash-strict-mode/