Add getRenderOptions utility function
This function checks markdown files for frontmatter with `gray-matter`.

If there is a "title" property in the frontmatter, it uses that as the title. Otherwise, it throws an error. This means that we need to make sure all our documentation markdown has frontmatter, so I've also added a test to check this on all the documentation markdown files.

This function can be used for other frontmatter properties in the future. For instance, I've added `caption` properties to the pages where we want to use them, but haven't yet implemented that in the function.

This shouldn't affect users, since these tests (and the function) target only our documentation site content. One impact is slightly slower completion time if a user runs `npm test`, because we don't yet separate our tests from user tests. We don't document this functionality, so it's likely only advanced users who are writing their own tests  would even be aware.