Use GitHub for update.sh instead of app files
We want to make sure that the update script is not included in the files
released with each version of the prototype kit, so we have an
`export-ignore` directive for that file (see edfa0362). However, this
also affects Heroku apparently (although it isn't documented), and
`update.sh` isn't included in the files in the 'slug' uploaded to be
deployed, which means we can't serve it. This is annoying, but as a
workaround we can instead direct user agents to GitHub.

Note that this means that users have to use the follow redirects flag
for curl, like so:

> curl -L https://govuk-prototype-kit.herokuapp.com/docs/update.sh