Fix update script tests in GitHub Actions
When running the test for the fetch step for the update script (which
accesses the GitHub Releases API) locally we haven't had any issues, but
when running it as part of the checks for a pull request some of the
checks fail with a `403 Forbidden` error. I suspect this is because
what the GitHub API sees is a sudden burst of traffic to one endpoint,
as a whole bunch of runners try to run the tests at the same time.

To avoid getting cut off we can use an access token. This PR changes the
test to use the automatic token [[1]] by adding the secret to the
environment for the job and then adding the authorization header to the
request. We make sure to only add the authorization header if the
GITHUB_TOKEN envvar is present, so that the tests can still be run
locally without an access token.

[1]: https://docs.github.com/en/actions/security-guides/automatic-token-authentication