Vendor gulp-nodemon 2.5.0
The package `gulp-nodemon` no longer appears to be maintained, which is
a problem, as we want it to remove its dependency on `colors`. We could
switch to a fork, or create our own fork, but as the amount of code
involved is small the simplest solution is to 'vendor' it: include it in
our codebase.

This commit adds a copy of the `index.js` file from the npm package
gulp-nodemon@2.5.0 (with an additional licence note prepended), and
changes the package manifest to require gulp-nodemons dependencies, but
not gulp-nodemon itself.

Note that we also add the vendored file to the list of files for
`standard` to ignore [[1]], so that the file doesn't need to deviate too
much from upstream (in case we want to check or merge changes at a later
date).

[1]: https://standardjs.com/#how-do-i-ignore-files