Treat Sass files from extensions as dependencies
Dart Sass allows you to silence deprecation warnings coming from dependencies (like GOV.UK Frontend, which is included via the extensions system) by using the `quietDeps` flag. A ‘dependency’ is defined as any file that’s loaded through `loadPaths` or `importer` [1]. Because we are using the legacy Javascript API, we use `includePaths` [2] which is equivalent to `loadPaths`.

In order to ensure that Sass files loaded by the extension system are treated as dependencies, we need to load them using this mechanism.

Add `node_modules` to the list of include paths when calling the Sass compiler, and modify the import paths generated by the `sass-extension` so that they are relative to `node_modules` rather than full file-system paths.

Update the two Sass files that manually import GOV.UK Frontend to import them relative to `node_modules`.

Enable the quietDeps flag and remove the silent logger as it is no longer needed.

[1]: https://sass-lang.com/documentation/js-api/interfaces/LegacySharedOptions#quietDeps
[2]: https://sass-lang.com/documentation/js-api/interfaces/LegacySharedOptions#includePaths