Add comment clarifying that Browsersync is only for local development
Browsersync doesn't play well with Glitch, so now it won't run if your
prototype is on Glitch, even if you have `useBrowserSync: true` in your
config file. I think this shouldn't actually affect users, because using
Browsersync with Glitch doesn't make much sense (it won't work, and
Glitch auto-relaods for you anyway), but this commit adds a comment
anyway to make it clear that Browsersync is only used for development
environments (the config variable has always been ignored when NODE_ENV
is true).