Do not install dev dependencies by default
Users of the prototype kit do not need to install dev dependencies such
as Jest or StandardJS [[1]]. Currently, because we distribute the kit as
a complete Node project rather than as a package, they are installing
them by default.

This commit restructures a few things to make sure that dev dependencies
are not installed unless explicitly requested. Unfortunately, the way we
need to to do this is a little convoluted.

The npm CLI does allow specifying that devDependencies should not be
installed by adding a config variable to the `.nvmrc` file [[2]];
however, if the directive `omit=dev` is given, which is what we want,
the NODE_ENV environment variable will also always be set to
`production`, which is definitely not what we want, as we use NODE_ENV
to determine whether a user is running the kit locally or on
Heroku/Glitch. So instead we move our devDependencies to be
optionalDependencies, say `omit=optional`, and tell developers to make
sure to ask for optional dependencies when installing (either with a
command line flag or the environment variable
`npm_config_include=optional`). This is an abuse of the
optionalDependencies feature, but feels less likely to break things for
users.

When/if the kit is made into a package, these changes can be reverted.

[1]: https://github.com/alphagov/govuk-prototype-kit/issues/1128
[2]: https://docs.npmjs.com/cli/v8/using-npm/config#omit