Allow integration tests to fail on macOS
Currently the prototype kit is flaky on macOS, `gulp.watch` doesn't
always pick up changes to the assets folder. We're planning to rip gulp
out soon (#1290), so for now we just let them fail, and we should just
make sure to fix the flakiness then so we can remove this kludge.