Remove Node.js unit test files from release archive
We have some unit tests which are saved as `.test.js` files next to the
implementation being tested. These are currently included in our release
archives, but they don't need to be. They won't be of interest to users.

This commit adds an ignore pattern to the gitattributes file, so that in
future these files won't be present at all.