Add caching for Cypress binary
Every time we run `npm ci`, the cypress package will try and download a
binary as a post-install step. We can save some bandwidth however
because it caches the binaries in `~/.cache/Cypress` [[1]]. This commit
adds a step to save the local cache to the GitHub Actions cache, which
should reduce the need for the extra downloads.

The code is based on the official example [[2]], except removing the lax
restore key to avoid cache 'snowballing' [[3]].

[1]: https://docs.cypress.io/guides/continuous-integration/introduction#Caching
[2]: https://github.com/cypress-io/cypress-example-kitchensink/blob/ab10094ef7b199ae7febafec413a0626414bcd3c/.github/workflows/single.yml
[3]: https://glebbahmutov.com/blog/do-not-let-cypress-cache-snowball/