Move checkFiles step to seperate node module
Split the step to check whether the node_modules folder is present to a
separate Node.js file. This makes it easier to ensure that the check can
run without any external dependencies required.

We then add a predev script to the package that calls checkFiles, to
make sure that checking whether the folder exists happens before
anything else.