Fix race condition in fetch test for update script
There is an unfortunate race condition in the test 'it downloads the
latest release of the prototype kit folder', because the update script
relies on the prototype kit docs app to direct it to the latest release
of the prototype kit, but the tests rely on the GitHub API for the same
information. When a release has been made on GitHub but the docs app
has not yet finished deployment, the tests will expect a newer version
of the kit than the update script will fetch. I couldn't really think of
a way around this, and besides I was already unhappy with the fact that
the tests were implicitly testing a function of the docs app, so instead
this commit loosens the requirements of the test slightly; now we just
assert that the script gets the latest release from the prototype kit
website.